# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/staff/login": {
      "post": {
        "operationId": "AuthController_staffLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Challenge issued"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Account suspended"
          }
        },
        "summary": "Step 1 of staff login: verify username/password, issue a 2FA challenge",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/staff/login/verify": {
      "post": {
        "operationId": "AuthController_verifyTwoFactor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyTwoFactorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login complete"
          },
          "401": {
            "description": "Invalid/expired code"
          }
        },
        "summary": "Step 2 of staff login: verify the 2FA code, issue tokens",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/agency/login": {
      "post": {
        "operationId": "AuthController_agencyLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgencyLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login complete"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Account or agency suspended"
          }
        },
        "summary": "Agency Portal login: phone+password, issues tokens directly (no 2FA)",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/agency/password-reset/request": {
      "post": {
        "operationId": "AuthController_agencyPasswordResetRequest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgencyPasswordResetRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "بازیابی رمز آژانس — درخواست OTP پیامکی",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/agency/password-reset/verify": {
      "post": {
        "operationId": "AuthController_agencyPasswordResetVerify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgencyPasswordResetVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "بازیابی رمز آژانس — تأیید OTP و صدور توکن",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/otp/request": {
      "post": {
        "operationId": "AuthController_requestOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Challenge issued"
          },
          "403": {
            "description": "Account suspended"
          }
        },
        "summary": "Public purchase engine — step 1: request an OTP for a customer phone number",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/otp/verify": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login complete"
          },
          "401": {
            "description": "Invalid/expired code"
          }
        },
        "summary": "Public purchase engine — step 2: verify the OTP, issue tokens",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/_test/last-otp/{phone}": {
      "get": {
        "operationId": "AuthController_testLastOtp",
        "parameters": [
          {
            "name": "phone",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "E2E-test only: reads back the mock OTP code. 404s in production.",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/me/locale": {
      "patch": {
        "operationId": "AuthController_updateLocale",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocaleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "ذخیرهٔ ترجیح زبان نمایش (fa/en/ar) — برای همگام‌سازی بین‌دستگاهی",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed"
          },
          "401": {
            "description": "Current password incorrect"
          }
        },
        "summary": "تغییر رمز عبور خود — رمز فعلی باید تأیید شود",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/set-password": {
      "post": {
        "operationId": "AuthController_setPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password set"
          }
        },
        "summary": "تعیین/بازنشانی رمز عبور مشتری پس از احراز هویت با OTP — بدون نیاز به رمز فعلی",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/customer/login-password": {
      "post": {
        "operationId": "AuthController_customerPasswordLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerPasswordLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login complete"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "ورود مشتری با موبایل + رمز عبور",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/password-reset/email/request": {
      "post": {
        "operationId": "AuthController_requestPasswordResetEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPasswordResetEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Challenge issued"
          },
          "401": {
            "description": "No account with a verified email matches"
          },
          "403": {
            "description": "Account suspended"
          }
        },
        "summary": "فراموشی رمز — مسیر ایمیل: ارسال کد به ایمیل تأییدشدهٔ حساب (جایگزین OTP پیامکی)",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/password-reset/email/verify": {
      "post": {
        "operationId": "AuthController_verifyPasswordResetEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPasswordResetEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login complete"
          },
          "401": {
            "description": "Invalid/expired code"
          }
        },
        "summary": "فراموشی رمز — مسیر ایمیل: تأیید کد، صدور توکن‌ها",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/_test/last-password-reset-email-code/{email}": {
      "get": {
        "operationId": "AuthController_testLastPasswordResetEmailCode",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "E2E-test only: reads back the mock password-reset-email code. 404s in production.",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/step-up/request": {
      "post": {
        "operationId": "AuthController_requestStepUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestStepUpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "درخواست کد تأیید مجدد پیش از عملیات پرریسک — تحویل از طریق کانال 2FA موجود",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/_test/last-code/{username}": {
      "get": {
        "operationId": "AuthController_testLastCode",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "E2E-test only: reads back the mock 2FA code. 404s in production.",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "blujet API",
    "description": "blujet airline platform — internal API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "StaffLoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "finance",
            "description": "Staff username"
          },
          "password": {
            "type": "string",
            "example": "Blujet@1404",
            "description": "Staff password"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "VerifyTwoFactorDto": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "example": "a1b2c3d4-...",
            "description": "Challenge id returned by /auth/staff/login"
          },
          "code": {
            "type": "string",
            "example": "482913",
            "description": "6-digit one-time code"
          }
        },
        "required": [
          "challengeId",
          "code"
        ]
      },
      "AgencyLoginDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+989120000002",
            "description": "شماره تماس آژانس"
          },
          "password": {
            "type": "string",
            "example": "Blujet@1404",
            "description": "رمز عبور"
          }
        },
        "required": [
          "phone",
          "password"
        ]
      },
      "AgencyPasswordResetRequestDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+989120000002"
          }
        },
        "required": [
          "phone"
        ]
      },
      "AgencyPasswordResetVerifyDto": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "challengeId",
          "code"
        ]
      },
      "RequestOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "09121234567",
            "description": "شماره موبایل مشتری"
          },
          "referralCode": {
            "type": "string",
            "example": "NEGAR-4152",
            "description": "کد معرف (فقط برای ثبت‌نام اول)"
          }
        },
        "required": [
          "phone"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "example": "a1b2c3d4-...",
            "description": "Challenge id returned by /auth/otp/request"
          },
          "code": {
            "type": "string",
            "example": "482913",
            "description": "6-digit one-time code"
          }
        },
        "required": [
          "challengeId",
          "code"
        ]
      },
      "UpdateLocaleDto": {
        "type": "object",
        "properties": {
          "locale": {
            "type": "string",
            "enum": [
              "FA",
              "EN",
              "AR"
            ],
            "example": "EN"
          }
        },
        "required": [
          "locale"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "رمز عبور فعلی"
          },
          "newPassword": {
            "type": "string",
            "description": "رمز عبور جدید — حداقل ۶ کاراکتر",
            "minLength": 6
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "SetPasswordDto": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string",
            "description": "رمز عبور جدید — حداقل ۸ کاراکتر با حروف بزرگ/کوچک، عدد و نماد",
            "minLength": 8,
            "example": "Blujet@1404"
          }
        },
        "required": [
          "newPassword"
        ]
      },
      "CustomerPasswordLoginDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "09121234567",
            "description": "شماره موبایل مشتری"
          },
          "password": {
            "type": "string",
            "description": "رمز عبور حساب"
          }
        },
        "required": [
          "phone",
          "password"
        ]
      },
      "RequestPasswordResetEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "negar@example.com",
            "description": "ایمیل تأییدشدهٔ حساب"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyPasswordResetEmailDto": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "example": "a1b2c3d4-...",
            "description": "Challenge id returned by /auth/password-reset/email/request"
          },
          "code": {
            "type": "string",
            "example": "482913",
            "description": "6-digit one-time code"
          }
        },
        "required": [
          "challengeId",
          "code"
        ]
      },
...
    }
  }
}